requirements.txt
Flask==3.0.3
flask-cors==4.0.1
flask-mysqldb==2.0.0

pandas==2.2.2

google-generativeai==0.7.2

python-dotenv==1.0.1

gunicorn==22.0.0